+++ NEWS.md
@@ -11,14 +11,21 @@
 * Added 'aspect' function
 * Added vignette with illustrations
 
-# starsExtra 0.1.0
+# starsExtra 0.1.0 (2020-06-25)
 
 * Added 'trim' function
 * Switched from 'testthat' to 'tinytest'
 * Replaced 'if' with 'switch' in C code
 * Added 'slope' function
 
+# starsExtra 0.1.2
+
+* Added multi-band sample data ('landsat')
+* Added 'st_normalize_2d' and 'st_normalize_3d' functions
+
 # To do
 
+* Adapt 'trim' to multi-band rasters
+* 3-band raster to 1-band RGB hex
 * Parallel processing in 'CI' function
-* Add 'extract' function (?exactextractr)
+* Add 'extract' function (exactextractr?)